Cheat Detector/requirements.txt
mlflow==2.9.2
numpy==1.24.3
tqdm==4.65.2
torchvision==0.16.0+cu121
torch==2.1.0+cu121
scikit-learn==1.3.2
matplotlib==3.7.2
fastapi==0.96.0
Pillow==10.0.1
uvicorn==0.27.0.post1
PyAutoGUI==0.9.54